test: remove IncidentWithFailingOperationZeebeImportIT
This test is not relevant for the new Camunda exporter case as it tests how
the import (export in our case) is behaving and retrying after operation
update failed. Operation update is processed differently now as a separate ExportHandlers,
comparing to update within Operate importing the incident that we had before.