fix: acknowledge retried correlate attempt
We need to acknowledge retried correlate attempts, because previous
acknowledgements may have been lost. If we don't ack, the subscription
may continue to retry correlating indefinitely.

The acknowledgement therefore needs to contain the message key that it
concerns.

However, the acknowledgement should be processed correctly. In some
cases we don't want to modify the state. We've implemented this in the
applier currently. However, we can also move this to the processor. This
would mean that the bug is fixed moving forward, and that we don't have
to make changes to the event applier. It also means that the events on
the log tell a better story.